moved model sources into tools folder
